Hex refactor
