--- wwe-api/settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name = 'com.wwe.wweapi'
-
