--- wwe-api/src/.gitignore
@@ -1,1 +1,0 @@
-/bin/
