--- wwe-api/src/main/java/com/wwe/wweapi/application/WweApiApplication.java
@@ -1,12 +1,0 @@
-package com.wwe.wweapi.application;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WweApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(WweApiApplication.class, args);
-	}
-}
