--- wwe-api/src/main/java/com/wwe/wweapi/domain/WrestlerService.java
@@ -1,11 +1,0 @@
-package com.wwe.wweapi.domain;
-
-import com.wwe.wweapi.domain.model.Wrestler;
-
-import java.util.List;
-
-public interface WrestlerService {
-
-    Wrestler getByID(String id);
-    List<Wrestler> getAll();
-}
