--- wwe-api/src/main/java/com/wwe/wweapi/domain/exception/WWEException.java
@@ -1,9 +1,0 @@
-package com.wwe.wweapi.domain.exception;
-
-public class WWEException extends Exception{
-
-    public WWEException(Throwable t){
-        super(t);
-    }
-
-}
