--- wwe-api/src/test/java/com/wwe/wweapi/WweApiApplicationTests.java
@@ -1,16 +1,0 @@
-package com.wwe.wweapi;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class WweApiApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
