+++ wweapi/src/main/java/com/wwe/api/wweapi/WweapiApplication.java
@@ -1,0 +1,12 @@
+package com.wwe.api.wweapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WweapiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WweapiApplication.class, args);
+	}
+}
