+++ wweapi/src/main/java/com/wwe/api/wweapi/domain/exception/WWEException.java
@@ -1,0 +1,9 @@
+package com.wwe.api.wweapi.domain.exception;
+
+public class WWEException extends Exception{
+
+    public WWEException(Throwable t){
+        super(t);
+    }
+
+}
