+++ wweapi/src/test/java/com/wwe/api/wweapi/WweapiApplicationTests.java
@@ -1,0 +1,16 @@
+package com.wwe.api.wweapi;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class WweapiApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
